+++ orcid-core/src/main/java/org/orcid/core/manager/v3/impl/OrcidSecurityManagerImpl.java
@@ -547,7 +547,7 @@ public class OrcidSecurityManagerImpl implements OrcidSecurityManager {
                     continue;
                 }
 
-                if (work.retrieveSourcePath().equals(clientId)) {
+                if (work.retrieveSourcePath() != null && work.retrieveSourcePath().equals(clientId)) {
                     filteredElements.add(work);
                     continue;
                 }