+++ orcid-web/src/main/java/org/orcid/frontend/web/controllers/PasswordResetController.java
@@ -24,6 +24,7 @@ import org.orcid.core.manager.ProfileEntityCacheManager;
 import org.orcid.core.manager.RegistrationManager;
 import org.orcid.core.manager.v3.ProfileEntityManager;
 import org.orcid.core.manager.v3.read_only.EmailManagerReadOnly;
+import org.orcid.core.togglz.Features;
 import org.orcid.core.utils.PasswordResetToken;
 import org.orcid.frontend.email.RecordEmailSender;
 import org.orcid.frontend.spring.ShibbolethAjaxAuthenticationSuccessHandler;
@@ -268,7 +269,6 @@ public class PasswordResetController extends BaseController {
         
         try {
              passwordResetToken = buildResetTokenFromEncryptedLink(oneTimeResetPasswordForm.getEncryptedEmail());
-
         } catch (EncryptionOperationNotPossibleException e) {
             oneTimeResetPasswordForm.getErrors().add("invalidPasswordResetToken");
             return oneTimeResetPasswordForm;
@@ -306,11 +306,18 @@ public class PasswordResetController extends BaseController {
         }
 
         profileEntityManager.updatePassword(orcid, oneTimeResetPasswordForm.getPassword().getValue());
+        //send the security notification email on change password
+        if(Features.SEND_EMAIL_ON_RESET_PASSWORD.isActive()) {
+			recordEmailSender.sendOrcidSecurityResetPasswordEmail(orcid);
+		}
         //reset the lock fields
         profileEntityManager.resetSigninLock(orcid);
         profileEntityCacheManager.remove(orcid);
         String redirectUrl = calculateRedirectUrl(request, response, false);
         oneTimeResetPasswordForm.setSuccessRedirectLocation(redirectUrl);
+        // Remove credentials before return
+        oneTimeResetPasswordForm.setPassword(null);
+        oneTimeResetPasswordForm.setRetypedPassword(null);
         return oneTimeResetPasswordForm;
     }